pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bordeaux-building-clustering"
version = "0.1.0"
description = "Clustering de typologies de bâtiments sur Bordeaux Métropole"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "geopandas>=0.14",
    "scikit-learn>=1.3",
    "hdbscan>=0.8",
    "umap-learn>=0.5",
    "folium>=0.15",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "esda>=2.5",
    "libpysal>=4.9",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "kneed>=0.8",
    "python-dotenv>=1.0",
    "pyarrow>=14.0",
    "statsmodels>=0.14",
    "joblib>=1.3",
    "shapely>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "jupyter>=1.0",
    "ipykernel>=6.0",
    "black>=23.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
